feat: optimize constructing the symbol table during write
The writer's internal symbol table is stored as hash map
from now on. This improves the runtime for determining
the index of a given symbol when constructing the index
table. In addition, the determination of the unique values,
during the construction of the symbol table, now uses a
dictionary. A dictionary has much better performance than
a simple for loop, but unlike a set, it also preserves the
order.